# docs/openapi/swagger.yaml
components:
  schemas:
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-array_github_com_Mininglamp-OSS_octo-marketplace_internal_service_category_AdminItem
    : properties:
        data:
          items:
            $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service_category.AdminItem'
          type: array
          uniqueItems: false
      type: object
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_api_response_EmptyResp
    : properties:
        data:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.EmptyResp'
      type: object
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_category_AdminItem
    : properties:
        data:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service_category.AdminItem'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.EmptyResp:
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error:
      properties:
        error:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody'
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        hint:
          type: string
        message:
          type: string
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_service_category.AdminItem:
      properties:
        icon_key:
          type: string
        name:
          type: string
        skill_category_id:
          type: string
        sort_order:
          type: integer
      type: object
...
    internal_api_handler_category.AdminCategoryRequest:
      properties:
        icon_key:
          type: string
        name:
          type: string
        sort_order:
          type: integer
      required:
      - name
      type: object
...
  securitySchemes:
    AdminToken:
      in: header
      name: X-Admin-Token
      type: apiKey
    Bearer:
      scheme: bearer
      type: http
info:
  contact:
    name: OCTO API Team
    url: https://github.com/Mininglamp-OSS/octo-marketplace
  description: Skill and MCP marketplace API for OCTO.
  title: Octo Marketplace API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /admin/skill_categories:
    get:
      description: Return non-deleted Skill categories for administrator management.
      operationId: skill_category.admin.list
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-array_github_com_Mininglamp-OSS_octo-marketplace_internal_service_category_AdminItem'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - AdminToken: []
      summary: List Skill categories (admin)
      tags:
      - skill_category
    post:
      description: Create a Skill category through the SuperAdmin-gated administrator
        surface.
      operationId: skill_category.create
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler_category.AdminCategoryRequest'
                description: Skill category
                summary: body
        description: Skill category
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_category_AdminItem'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: CONFLICT
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - AdminToken: []
      summary: Create Skill category
      tags:
      - skill_category
  /admin/skill_categories/{skill_category_id}:
    delete:
      description: Delete an unused Skill category through the SuperAdmin-gated administrator
        surface; categories referenced by Skills are rejected.
      operationId: skill_category.delete
      parameters:
      - description: Skill category ID
        in: path
        name: skill_category_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_api_response_EmptyResp'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: CONFLICT
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - AdminToken: []
      summary: Delete Skill category
      tags:
      - skill_category
    patch:
      description: Replace mutable fields of an existing Skill category through the
        SuperAdmin-gated administrator surface.
      operationId: skill_category.update
      parameters:
      - description: Skill category ID
        in: path
        name: skill_category_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler_category.AdminCategoryRequest'
                description: Skill category changes
                summary: body
        description: Skill category changes
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_category_AdminItem'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - AdminToken: []
      summary: Update Skill category
      tags:
      - skill_category
...
servers:
- url: /v1
...